+++ src/conversation.js
@@ -277,6 +277,7 @@ class Conversation extends Syncable {
       participants: this.participants,
       distinct: this.distinct,
       metadata: isMetadataEmpty ? null : this.metadata,
+      id: this.id,
     };
   }
 
@@ -287,7 +288,7 @@ class Conversation extends Syncable {
    * events reporting changes to the layer.Conversation.id can
    * be applied before reporting on it being sent.
    *
-   * Example: Query will now have IDs rather than TEMP_IDs
+   * Example: Query will now have the resolved Distinct IDs rather than the proposed ID
    * when this event is triggered.
    *
    * @method _createResult
@@ -354,8 +355,9 @@ class Conversation extends Syncable {
 
     const id = this.id;
     this.id = conversation.id;
+
+    // IDs change if the server returns a matching Distinct Conversation
     if (id !== this.id) {
-      this._tempId = id;
       client._updateConversationId(this, id);
       this._triggerAsync('conversations:change', {
         oldValue: id,
@@ -939,7 +941,7 @@ class Conversation extends Syncable {
     if (!result.success) {
       this.syncState = Constants.SYNC_STATE.NEW;
       this.trigger('conversations:loaded-error', { error: result.data });
-      this.destroy();
+      if (!this.isDestroyed) this.destroy();
     } else {
       // If successful, copy the properties into this object
       this._populateFromServer(result.data);
@@ -1416,18 +1418,6 @@ Conversation.prototype._toObject = null;
  * @private
  */
 Conversation.prototype._sendDistinctEvent = null;
-
-/**
- * A locally created Conversation will get a temporary ID.
- *
- * Some may try to lookup the Conversation using the temporary ID even
- * though it may have later received an ID from the server.
- * Keep the temporary ID so we can correctly index and cleanup.
- *
- * @type {String}
- * @private
- */
-Conversation.prototype._tempId = '';
 
 /**
  * Prefix to use when generating an ID for instances of this class