+++ test/specs/unit/clientSpec.js
@@ -246,11 +246,6 @@ describe("The Client class", function() {
             }).toThrowError(layer.LayerError.dictionary.idParamRequired);
             expect(layer.LayerError.dictionary.idParamRequired.length > 0).toBe(true);
         });
-
-        it("Should work with a tempId", function() {
-          client._tempConversationsHash["temp"] = conversation.id;
-          expect(client.getConversation("temp")).toBe(conversation);
-        });
     });
 
     describe("The _addConversation() method", function() {
@@ -344,20 +339,6 @@ describe("The Client class", function() {
             expect(client._conversationsHash).toEqual(hash);
         });
 
-        it("Should delete from _tempConversationsHash", function() {
-            // Setup
-            var c1 = client.createConversation(["a"]);
-            c1._tempId = "temp";
-            client._tempConversationsHash[c1._tempId] = c1.id;
-
-
-            // Run
-            client._removeConversation(c1);
-
-            // Posttest
-            expect(client._tempConversationsHash["temp"]).toBe(undefined);
-        });
-
         it("Should trigger event on removing conversation", function() {
             // Setup
             var c1 = new layer.Conversation({});
@@ -444,20 +425,6 @@ describe("The Client class", function() {
             expect(client._conversationsHash[c1id]).toBe(undefined);
         });
 
-        it("Should enter it into _tempConversationsHash", function() {
-            // Setup
-            var c1 = new layer.Conversation({});
-            client._addConversation(c1);
-            var c1id = c1.id;
-
-            // Run
-            c1.id = "fred";
-            client._updateConversationId(c1, c1id);
-
-            // Posttest
-            expect(client._tempConversationsHash[c1id]).toEqual("fred");
-        });
-
         it("Should update all Message conversationIds", function() {
             // Setup
             var c1 = new layer.Conversation({participants: ["a"]});
@@ -515,7 +482,7 @@ describe("The Client class", function() {
         });
 
         it("Should load by id", function() {
-            var newId = message.id.replace(/temp_/,"") + "a";
+            var newId = message.id + "a";
             var m1 = client.getMessage(newId, true);
 
             // Posttest
@@ -529,11 +496,6 @@ describe("The Client class", function() {
                 client.getMessage(5);
             }).toThrowError(layer.LayerError.dictionary.idParamRequired);
             expect(layer.LayerError.dictionary.idParamRequired.length > 0).toBe(true);
-        });
-
-        it("Should work with a tempId", function() {
-          client._tempMessagesHash["temp"] = message.id;
-          expect(client.getMessage("temp")).toBe(message);
         });
     });
 
@@ -807,7 +769,6 @@ describe("The Client class", function() {
             // Posttest
             expect(client._tempMessagesHash[mId]).toEqual("fred");
         });
-
     });
 
     describe("The _getObject() method", function() {