+++ test/specs/unit/conversationSpec.js
@@ -458,6 +458,10 @@ describe("The Conversation Class", function() {
             conversation.metadata = {a: "b", c: "d"};
             expect(conversation._getPostData().metadata).toEqual({a: "b", c: "d"});
         });
+
+        it("Should return the Conversation ID", function() {
+          expect(conversation._getPostData().id).toEqual(conversation.id);
+        });
     });
 
     describe("The _createResult() method", function() {
@@ -764,19 +768,6 @@ describe("The Conversation Class", function() {
                     newValue: conversation.id,
                     property: 'id',
                 });
-        });
-
-        it("Should write a _tempId property", function() {
-            // Setup
-            spyOn(conversation, "_triggerAsync");
-            var initialId = conversation.id;
-
-            // Run
-            conversation._populateFromServer(c);
-
-            // Posttest
-            expect(conversation._tempId).toEqual(initialId);
-            expect(conversation.id).not.toEqual(initialId);
         });
 
         it("Should setup lastMessage", function() {